Refactor Horizon module to package structure under apts/horizon/
