+++ apts/horizon/__init__.py
@@ -1,0 +1,3 @@
+from .base import Horizon
+
+__all__ = ["Horizon"]
